Protect score export generation from dividing by zero
A cvcn experiment exited because there were no subjects in the real
threshold range, causing a division by 0. The gold labels weren't
being logged anymore, and the trial was not uploaded to mongo, meaning
there is no way to reproduce the edge case and determine what happened.
This will allow the process to continue and upload meaningful data
to determine what happened (if it happens again)